requirements.txt
fastapi==0.99.3
uvicorn[standard]==0.21.1
SQLAlchemy==2.2.3
psycopg2-binary==2.9.6
alembic==1.11.1
pydantic==1.10.7
python-multipart==0.0.6
boto3==1.28.46
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.27.0
redis==4.7.2
